Share the jsx-runtime external across webpack configs
The defensive react/jsx-runtime external string was duplicated verbatim
in the renderer, dcc, html and table webpack configs. Extract it into a
single module (dash-renderer/jsx-runtime-external.js), co-located with
react-shim.js which it must stay in sync with, and require it from all
four configs.
