+++ components/dash-table/.config/webpack/base.js
@@ -3,48 +3,11 @@ const WebpackDashDynamicImport = require('@plotly/webpack-dash-dynamic-import');
 
 const basePreprocessing = require('./base.preprocessing');
 const packagejson = require('./../../package.json');
+const {
+    jsxRuntimeExternal
+} = require('../../../../dash/dash-renderer/jsx-runtime-external');
 
 const dashLibraryName = packagejson.name.replace(/-/g, '_');
-
-// Externalized react/jsx-runtime. Newer Dash provides window.ReactJSXRuntime
-// (see dash-renderer/src/react-shim.js) backed by the React version loaded on
-// the page; the inline fallback keeps this bundle working on older Dash,
-// which does not define the global. Keep in sync with react-shim.js.
-const jsxRuntimeExternal = `var (window.ReactJSXRuntime || (window.ReactJSXRuntime = (function (React) {
-    function jsx(type, config, maybeKey) {
-        var props = {};
-        var children = null;
-        if (config != null) {
-            if (config.key !== undefined) {
-                props.key = '' + config.key;
-            }
-            for (var propName in config) {
-                if (
-                    Object.prototype.hasOwnProperty.call(config, propName) &&
-                    propName !== 'key' &&
-                    propName !== '__self' &&
-                    propName !== '__source'
-                ) {
-                    if (propName === 'children') {
-                        children = config[propName];
-                    } else {
-                        props[propName] = config[propName];
-                    }
-                }
-            }
-        }
-        if (maybeKey !== undefined) {
-            props.key = '' + maybeKey;
-        }
-        if (children === null || children === undefined) {
-            return React.createElement(type, props);
-        }
-        return Array.isArray(children)
-            ? React.createElement.apply(React, [type, props].concat(children))
-            : React.createElement(type, props, children);
-    }
-    return {jsx: jsx, jsxs: jsx, jsxDEV: jsx, Fragment: React.Fragment};
-})(window.React)))`;
 
 
 module.exports = (options = {}) => {
